requirements.txt
base32hex==1.0.2
tinyec==0.4.0
socket
pycryptodome==3.16.0
secrets
pyjwt[crypto]
openpyxl==3.0.10
PySimpleGUI==4.60.4
wheel==0.38.4
setuptools==65.6.3